Changes interaction animation with message and custom enum
