+++ Assets/_Core/Animations/Animator/PlayerAnimator.controller
@@ -166,7 +166,7 @@ AnimatorState:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: PlayerInteractDesactiveInput
+  m_Name: PlayerInteractDisableInput
   m_Speed: 1
   m_CycleOffset: 0
   m_Transitions:
