+++ Assets/_Core/Messages/String.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7e49fac4725d4944cb9be577b8db8c58
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
