+++ Assets/_Core/Messages/String/AnimationMessage.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6405500acd090ce49a878021fd371e5e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
