+++ Assets/_Core/Scripts/Animations/EAnimations.cs
@@ -1,0 +1,8 @@
+﻿namespace AlreadyGone.Animations
+{
+    public enum EAnimations
+    {
+        PlayerInteractActiveInput,
+        PlayerInteractDisableInput,
+    }
+}
