+++ Assets/_Core/Scripts/Animations/EAnimations.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: d795ee2f9e3e4a93aa8642d04544094e
+timeCreated: 1749032601
