+++ Assets/_Core/Scripts/DesignPattern/Observers/Typed/EAnimation.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 9c7a88d3742744cebd42abaa6c98af7f
+timeCreated: 1749032217
