+++ Assets/_Core/Scripts/DesignPattern/Observers/Typed/EAnimation/MessageEAnimation.cs
@@ -1,0 +1,8 @@
+﻿using UnityEngine;
+
+namespace AlreadyGone.DesignPattern.Observers.Typed.EAnimation
+{
+    [CreateAssetMenu(fileName = "new_" + nameof(MessageEAnimation), menuName = "Messages/String")]
+    public class MessageEAnimation : TypedMessage<Animations.EAnimations>
+    { }
+}
