+++ Assets/_Core/Scripts/DesignPattern/Observers/Typed/EAnimation/MessageEAnimation.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: f6c525d8d6484c1ab94b91461c50e9b2
+timeCreated: 1749032229
