+++ Assets/_Core/Scripts/DesignPattern/Observers/Typed/EAnimation/MessageListenerEAnimation.cs
@@ -1,0 +1,5 @@
+﻿namespace AlreadyGone.DesignPattern.Observers.Typed.EAnimation
+{
+    public class MessageListenerEAnimation :  TypedMessageListener<Animations.EAnimations>
+    { }
+}
