+++ Assets/_Core/Scripts/DesignPattern/Observers/Typed/EAnimation/MessageListenerEAnimation.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 9707eef4262f450dacd442446437071a
+timeCreated: 1749032229
