+++ Assets/_Core/Scripts/Interactions/IInteractable.cs
@@ -1,7 +1,10 @@
 namespace AlreadyGone.Interactions
 {
+    using Animations;
+    
     public interface IInteractable
     {
+        public EAnimations GetAnimation();
         public void VisualizeInteraction();
         public void HideInteraction();
         public void Interact();
