+++ Assets/_Core/Scripts/Interactions/Interactable/InteractionTestWithColor.cs.meta
